Fix issues with headers not included from ProviderRdtsc.
