+++ src/apps/refactor_proto/refactor_proto/CMakeLists.txt
@@ -46,8 +46,9 @@ set(${TARGET_NAME}_SOURCES
     SceneGui.cpp
     SetupDrawing.cpp
 
-    providers/ProviderGL.cpp
     providers/ProviderCpu.cpp
+    providers/ProviderGL.cpp
+    providers/ProviderRdtsc.cpp
 )
