+++ src/apps/refactor_proto/refactor_proto/providers/ProviderRdtsc.h
@@ -4,8 +4,22 @@
 #include "../Profiler.h"
 #include "../Time.h"
 
+#ifdef _WIN32
+#include <intrin.h>
+#endif
+
 
 namespace ad::renderer {
+
+
+inline std::uint64_t readTsc()
+{
+    unsigned int ui;
+    return __rdtscp(&ui);
+}
+
+
+Ratio evluateRtscTickPeriodMicroSeconds();
 
 
 struct ProviderCpuRdtsc : public ProviderInterface
@@ -29,37 +43,6 @@ struct ProviderCpuRdtsc : public ProviderInterface
     std::vector<TickCount_t> mIntervals;
 };
 
-
-inline std::uint64_t readTsc()
-{
-    unsigned int ui;
-    return __rdtscp(&ui);
-}
-
-
-Ratio evluateRtscTickPeriodMicroSeconds()
-{
-    auto clockStart = Clock::now();
-    auto tickCount = readTsc();
-
-    std::this_thread::sleep_for(std::chrono::milliseconds{100});
-
-    auto clockDuration = Clock::now() - clockStart;
-    tickCount = readTsc() - tickCount;
-
-    std::uint64_t ticksPerMicrosecond =
-        tickCount / getTicks<std::chrono::microseconds>(clockDuration);
-    SELOG(info)("TSC frequency evaluated to: {} MHz", ticksPerMicrosecond);
-
-    // Check it does not overflow
-    //assert(ticksPerMicrosecond < std::numeric_limits<Sample_t>::max());
-    static_assert(std::is_same_v<ProviderInterface::Sample_t, std::uint64_t>);
-
-    return Ratio{
-        .num = 1,
-        .den = ticksPerMicrosecond,
-    };
-}
 
 //
 // Implementations
@@ -88,7 +71,7 @@ inline void ProviderCpuRdtsc::endSection(EntryIndex aEntryIndex, std::uint32_t a
 }
 
 
-bool ProviderCpuRdtsc::provide(EntryIndex aEntryIndex, uint32_t aQueryFrame, Sample_t & aSampleResult)
+inline bool ProviderCpuRdtsc::provide(EntryIndex aEntryIndex, uint32_t aQueryFrame, Sample_t & aSampleResult)
 {
     const auto & interval = getInterval(aEntryIndex, aQueryFrame);
     aSampleResult = interval;
